PromptEngineering/python/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
beautifulsoup4
youtube-transcript-api
